CU pedido

+++ database/migrations/2023_05_23_183637_create_ingredientes_table.php
@@ -15,10 +15,10 @@ return new class extends Migration
             $table->id();
             $table->string('nombre', 150)->unique();
             $table->string('unidad', 100);
-            $table->integer('stock')->default(0);
+            $table->decimal('stock', 10, 2)->default(0);
             $table->decimal('precio_unidad', 10, 2)->default(0.00);
-            $table->integer('stock_minimo')->default(0);
-            $table->integer('stock_maximo')->default(0);
+            $table->decimal('stock_minimo', 10, 2)->default(0);
+            $table->decimal('stock_maximo', 10, 2)->default(0);
             $table->text('descripcion')->nullable();
             $table->timestamps();
         });
