+++ database/migrations/2023_05_23_183738_create_preparacions_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
             $table->id();
             $table->foreignId('receta_id')->constrained('recetas')->onDelete('cascade');
             $table->foreignId('ingrediente_id')->constrained('ingredientes')->onDelete('cascade');
-            $table->integer('cantidad')->default(0);
+            $table->decimal('cantidad', 10, 2)->default(0.00);
             $table->decimal('costo', 10, 2)->default(0.00);
             $table->timestamps();
         });
